requirements.txt
PyYAML==6.0
requests==2.28.1
flask==2.2.2
python-dotenv
typing
